% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_demo.R
\name{dl_demo}
\alias{dl_demo}
\title{Demo data downloader}
\usage{
dl_demo(path = tempdir(), quiet = FALSE)
}
\arguments{
\item{path}{Path to which the demo data folder will be downloaded. Default is tempdir().}

\item{quiet}{A boolean stating whether to print the progress and outcome of the downloading. Default is FALSE.}
}
\value{
A string containing the path to the downloaded demo_data folder
}
\description{
Downloads all demo data used in the package's examples and vignettes at once.
}
\examples{
dl_demo()
}
