% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_TFCE.R
\name{LM_TFCE}
\alias{LM_TFCE}
\title{Voxel-wise linear model analysis with threshold-free cluster enhancement}
\usage{
LM_TFCE(
  model,
  contrast,
  formula,
  formula_dataset,
  vox_data,
  coords,
  smooth_FWHM,
  nperm = 1000,
  tail = 2,
  nthread = 4
)
}
\arguments{
\item{model}{An N X P data.frame object containing N rows for each subject and P columns for each predictor included in the model.This data.frame should not include the random effects variable.}

\item{contrast}{A N x 1 numeric vector or object containing the values of the predictor of interest. Its length should equal the number of subjects in model (and can be a single column from model). The t-stat and TFCE maps will be estimated only for this predictor.}

\item{formula}{An optional string or formula object describing the predictors to be fitted against the surface data, replacing the model, contrast, or random arguments. If this argument is used, the formula_dataset argument must also be provided.
\itemize{
\item The dependent variable (DV) is not needed, and the formula will start with ~. The DV will be the surface data value by default, but it can be swapped with contrast as IV via the "inverse" argument.
\item The first independent variable in the formula will always be interpreted as the contrast of interest for which to estimate cluster-thresholded t-stat maps.
\item Only one random regressor can be given and must be indicated as '(1|variable_name)'.
}}

\item{formula_dataset}{An optional data.frame object containing the independent variables to be used with the formula (the IV names in the formula must match their column names in the dataset).}

\item{vox_data}{A N x V matrix object containing the voxel-wse data (N row for each subject, V for each voxel)}

\item{coords}{A V x 3 matrix, containing the X, Y and Z coordinates of the voxels used in \code{vox_data}.}

\item{smooth_FWHM}{A numeric value specifying the desired smoothing width in mm. It should not be specified if the vox_data has been smoothed previously with smooth_vox(), because this results in vox_data being smoothed twice.}

\item{nperm}{An integer specifying the number of permutations generated for the subsequent thresholding procedures (default = 100)}

\item{tail}{An integer specifying whether to test a one-sided positive (1), one-sided negative (-1) or two-sided (2) hypothesis}

\item{nthread}{An integer specifying the number of CPU threads to allocate}
}
\value{
A list object containing the t-stat and the TFCE statistical maps which can then be subsequently thresholded using TFCE_threshold()
}
\description{
Fits a linear model with the voxel-wise data as the predicted outcome, and returns t-stat and threshold-free cluster enhancement (TFCE) statistical maps for the selected contrast.
}
\details{
This TFCE method is adapted from the \href{https://github.com/nilearn/nilearn/blob/main/nilearn/mass_univariate/_utils.py#L7C8-L7C8}{'Nilearn' Python library}.
}
