% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsi_extract.R
\name{qsi_extract}
\alias{qsi_extract}
\title{QSI diffusion-weighted imaging metrics extractor}
\usage{
qsi_extract(
  inputdir,
  outputdir,
  metrics = c("fa", "md"),
  skeleton_fathreshold = 0.2,
  dti_tensor = "dtiTensor",
  dti_method,
  dti_sigma = NULL,
  dti_L = 1,
  nthread = 4,
  keep_maps = FALSE,
  qsiprep_path = NULL,
  silent = FALSE
)
}
\arguments{
\item{inputdir}{A string object containing the path to the QSIprep or QSIrecon output dataset. For QSIrecon, specify "derivatives/qsirecon-*/" instead of the parent directory, as some files have identical suffixes and cannot be disentangled across reconstructions.}

\item{outputdir}{A string object containing the path of the directory where the final cohort-wise RDS will be stored for each metric (as well as metrics maps if \code{keep_maps} is set as TRUE). Default is 'cohort_skeletons' in the R temporary directory (tempdir()).}

\item{metrics}{A string object or vector of string objects containing the name (s) of the metric(s) to be estimated, in lower case. For QSIprep outputs, only the dti package's "dtiIndices/dkiIndices" metrics apply ("fa","ga","md","k1","k2","k3","mk","mk2","kaxial","kradial","fak"); for QSIrecon outputs, it can be any reconstruction output that has a "\emph{param-\}_dwimap.nii.gz" suffix in MNI152 space (e.g., 'icvf', 'od' etc.). Default is c('fa', 'md').}

\item{skeleton_fathreshold}{A numerical object with the Fractional Anisotropy (FA) threshold value to apply on the template FA skeleton (FMRIB58 2mm). Default is 0.2.}

\item{dti_tensor}{A string object stating the tensor to be used for applicable metrics estimation ('dtiTensor' or 'dkiTensor'). Default is dtiTensor. Argument ignored for QSIrecon output.}

\item{dti_method}{A string object containing the method to be used for tensor-based estimations. If \code{dti_tensor} is 'dtiTensor', options include "nonlinear", "linear" (default), "quasi-likelihood"; if \code{dti_tensor} is 'dkiTensor', options include "CLLS-QP" (default), "CLLS-H", "ULLS", "QL", "NLR". Argument ignored for QSIrecon output.}

\item{dti_sigma}{An integer specifying the sigma value (scale parameter of the signal's distribution) to be used as part of the tensor estimation. Default is NULL. Argument ignored for QSIrecon output.}

\item{dti_L}{An integer specifying the effective degrees of freedom for the tensor estimation. Default is 1.  Argument ignored for QSIrecon output.}

\item{nthread}{Number of CPU threads for tensor estimation and ITK work units for image resampling.}

\item{keep_maps}{A logical object to determine whether files such as estimated tensor maps and coregistered maps are to be written in the \code{outputdir}. Default is FALSE.}

\item{qsiprep_path}{A string containing the path to the QSIprep output (Optional). Ignored if inputdir and qsiprep_path are the same. Its purpose is for QSIrecon processing to retrieve ACPC-to-MNI152 transformation matrices if MNI152 coregistration was not done by QSIrecon. The QSIprep folder must have the same subjects and sessions as the the QSIrecon output.}

\item{silent}{A logical object to determine whether messages will be silenced. Default is FALSE.}
}
\value{
A list of 2D matrices, each matrix corresponding to one metric from
\code{metrics}. Each element (skel_matrices$fa, skel_matrices$md, skel_matrices$ga, ...) is its own separate matrix: rows = subjects (and sessions), columns = voxels. Additionally, the list contains the coordinates of the skeleton voxels (skel_coords matrix), the skeleton template they are based on, and the FA threshold selected.
}
\description{
Extracts diffusion weighted imaging based metrics across a whole cohort datasets from QSIprep or QSIrecon pipeline outputs, masked into a common skeleton template, and merging them into single RDS files for each metric of interest.
}
\details{
For QSIprep output, the function makes use of tools from the \code{dti} R package to build a diffusion-weighted map and estimate DTI or DKI tensors. This requires bvec, bval and the dwi images to all be present in the subject directory. The map estimated from it is then coregistered to MNI 152 2mm space using the native ITK C++ backend and the ACPC-to-MNI152 transforms that QSIprep generates.
For QSIrecon outputs, maps are already present in MNI 152 space, and the qsi_extract() function regrids MNI152 maps whenever their physical grid differs from the skeleton template.
The FA skeleton is based on FSL's FMRIB58_FA-skeleton_1mm downsampled to 2mm.
}
