% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsi_native.R
\name{qsi_image_write}
\alias{qsi_image_write}
\title{Write a scalar 3-D NIfTI image using RNifti}
\usage{
qsi_image_write(image, filename, singleprecision = TRUE)
}
\arguments{
\item{image}{NIfTI filename or RNifti niftiImage.}

\item{filename}{Output .nii or .nii.gz filename.}

\item{singleprecision}{Write float32 rather than float64.}
}
\value{
Output filename, invisibly.
}
\description{
Write a scalar 3-D NIfTI image using RNifti
}
