% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsi_native.R
\name{qsi_resample_to_target}
\alias{qsi_resample_to_target}
\title{Resample a scalar image to a reference grid}
\usage{
qsi_resample_to_target(moving, fixed, interpolator = "linear", nthread = 1L)
}
\arguments{
\item{moving, fixed}{NIfTI filename or RNifti niftiImage.}

\item{interpolator}{"linear" or "nearestNeighbor".}

\item{nthread}{Positive number of threads requested for the ANTs subprocess.}
}
\value{
RNifti niftiImage.
}
\description{
Resample a scalar image to a reference grid
}
