% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_vox}
\alias{smooth_vox}
\title{3D Spatial Smoothing for Voxel Matrices}
\usage{
smooth_vox(
  data_mat,
  coords,
  fwhm = NULL,
  sigma = NULL,
  voxel_size = c(1, 1, 1)
)
}
\arguments{
\item{data_mat}{Numeric matrix of size N x V (N subjects, V voxels).}

\item{coords}{Numeric or integer matrix of size V x 3 containing X, Y, Z voxel grid coordinates.}

\item{fwhm}{Full Width at Half Maximum (scalar or length-3 vector in mm).}

\item{sigma}{Gaussian standard deviation (scalar or length-3 vector in mm). Ignored if \code{fwhm} is given.}

\item{voxel_size}{Voxel dimension in mm (scalar or length-3 vector). Default is c(1, 1, 1).}
}
\value{
A smoothed N x V matrix.
}
\description{
3D Spatial Smoothing for Voxel Matrices
}
