% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFCE_threshold.R
\name{TFCE_threshold}
\alias{TFCE_threshold}
\title{Thresholding TFCE output}
\usage{
TFCE_threshold(TFCEoutput, p = 0.05, k = 20)
}
\arguments{
\item{TFCEoutput}{An object containing the output from TFCE_vertex_analysis()}

\item{p}{A numeric object specifying the p-value to threshold the results (Default is 0.05)}

\item{k}{Cluster-forming threshold (Default is 20)}

\item{atlas}{An integer corresponding to the atlas of interest.  1=Desikan, 2=Destrieux-148, 3=Glasser-360, 4=Schaefer-100, 5=Schaefer-200, 6=Schaefer-400. Set to \code{1} by default. This argument is ignored for hippocampal surfaces. For applicable SubCortexMesh surfaces, 1=default base ROI, 2=anatomical atlas.}
}
\value{
A list object containing the cluster level results, unthresholded t-stat map, thresholded t-stat map, and positive, negative and bidirectional cluster maps.
}
\description{
Threshold TFCE maps from the TFCE_vertex_analysis() output and identifies significant clusters at the desired threshold.
}
