% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsi_extract.R
\name{qsi_extract}
\alias{qsi_extract}
\title{QSI diffusion-weighted imaging metrics extractor}
\usage{
qsi_extract(
  inputdir,
  outputdir,
  metrics = c("fa", "md", "ga"),
  skeleton_threshold = 0.2,
  dti_tensor = "dtiTensor",
  dti_method,
  dti_sigma = NULL,
  dti_L = 1,
  nthread = 4,
  keep_maps = FALSE,
  silent = FALSE
)
}
\arguments{
\item{inputdir}{A string object containing the path to the QSIprep or QSIrecon output dataset. For QSIrecon, specify "derivatives/qsirecon-*/" instead of the parent directory, as some files have identical suffixes and cannot be disentangled across reconstructions.}

\item{outputdir}{A string object containing the path of the directory where the final cohort-wise RDS will be stored for each metric (as well as metrics maps if \code{keep_maps} is set as TRUE). Default is 'cohort_skeletons' in the R temporary directory (tempdir()).}

\item{metrics}{A string object or vector of string objects containing the name (s) of the metrics to be estimated. For QSIprep outputs, only the dti package's "dtiIndices/dkiIndices" metrics apply ("fa","ga","md","k1","k2","k3","mk","mk2","kaxial","kradial","fak"); for QSIrecon outputs, it can be any reconstruction output that has a "\emph{param-\}_dwimap.nii.gz" output in MNI152 space. Default is 0.2.}

\item{skeleton_threshold}{A numerical object with the (Fractional Anisotropy) threshold value with which to apply the template skeleton.}

\item{dti_tensor}{A string object stating the tensor to be used for applicable metrics estimation ('dtiTensor' or 'dkiTensor'). Default is dtiTensor. Argument ignored for QSIrecon output.}

\item{dti_method}{A string object containing the method to be used for tensor-based estimations. If \code{dti_tensor} is dtiTensor, it includes "nonlinear" (default), "linear", "quasi-likelihood"; if \code{dti_tensor} is dkiTensor, it includes "CLLS-QP" (default), "CLLS-H", "ULLS", "QL", "NLR". Argument ignored for QSIrecon output.}

\item{dti_sigma}{An integer specifying the sigma value (scale parameter of the signal's distribution) to be used as part of the tensor estimation. Default is NULL. Argument ignored for QSIrecon output.}

\item{dti_L}{An integer specifying the effective degrees of freedom. Default is 1.  Argument ignored for QSIrecon output.}

\item{nthread}{Number of CPU threads for the dti package to use when estimating the tensor and metrics from the data. Argument ignored for QSIrecon output.}

\item{keep_maps}{A logical object to determine whether files such as estimated tensor maps and coregistered maps are to be written in the \code{outputdir}. Default is FALSE.}

\item{silent}{A logical object to determine whether messages will be silenced. Default is FALSE.}
}
\value{
A list of 2D matrices, each matrix corresponding to one metric from
\code{metrics}. Each element (skel_matrices$fa, skel_matrices$md, skel_matrices$ga, ...) is its own separate matrix: rows = sub_ses, columns = voxels. Additionally
}
\description{
Extracts diffusion weighted imaging based metrics across a whole cohort datasets from QSIprep or QSIrecon pipeline outputs, masked into a common skeleton template, and merging them into single RDS files for each metric of interest.
}
\details{
For QSIprep output, the function makes use of tools from the \code{dti} R package to build a diffusion-weighted map and estimate DTI or DKI tensors. This requires bvec, bval and the dwi images to all be present in the subject directory. The map estimated from it is then coregistered to MNI 152 2mm space using the \code{rpyANTs} package and the ACPC-to-MNI152 transforms that QSIprep generates.
For QSIrecon outputs, maps are already present in MNI 152 space, and the qsi_extract() function will only regrid the MNI152 maps to 2 mm if resolution is higher.
The FA skeleton is based on FSL's FMRIB58_FA-skeleton_1mm downsampled to 2mm.
}
\examples{
SCMvextract(sdirpath = "subcortexmesh_output_metrics", 
outputdir=paste0(tempdir(), "\\\\subcortices"), template='fsaverage', measure="surfarea") 
}
